refactor(validation): lift downstream-liveness wiring into the library
The per-architecture downstream (fall-through) liveness scanners lived in
the `main.rs` binary root: `fall_through_suffix`, the AArch64/x86 Capstone
decoders, four `*_from_bytes` scanners, and four `*_from_section` wrappers.
The shared scanning discipline (`scan_flags_live`/`scan_regs_live`) already
lived in `validation::downstream`, so the wiring was split across the
binary/library boundary and its ~15 characterization tests could only run
through the binary's in-file test module.

Move the wiring next to the discipline it builds on. `fall_through_suffix`
and the two Capstone→IR decoders become private module internals; the byte
scanners stay module-private; and the module now exposes one narrow
section-level entry point per (architecture, scanned state):
`aarch64_downstream_flags_live`, `aarch64_downstream_regs_live`,
`x86_downstream_flags_live`, `x86_downstream_regs_live`. `main.rs` drops
~560 lines and calls the four public entries; the soundness default (an
un-analyzable suffix keeps everything live) now has one owner.

Behavior-preserving. Adds direct section-level unit tests for the new
public interface, migrates the byte-level scanner tests into the module,
and adds a shared minimal-ELF builder to `test_utils`.
