+++ src/validation/downstream.rs
@@ -13,10 +13,16 @@
 //!   (the AArch64 Capstone bridge vs. `x86_ir_from_mnemonic`), and
 //! * the per-instruction liveness rule applied to a decoded instruction.
 //!
-//! Both knobs are supplied by the caller as closures, so this module owns the
-//! shared scanning discipline once — including the soundness default that an
-//! *un-analyzable* suffix keeps everything live — instead of the four
-//! copy-pasted scanners the CLI used to carry.
+//! [`scan_flags_live`] and [`scan_regs_live`] own the shared scanning discipline
+//! once — including the soundness default that an *un-analyzable* suffix keeps
+//! everything live — taking those two knobs as closures. On top of them this
+//! module supplies the per-architecture wiring (the Capstone→IR decoders and the
+//! fall-through suffix resolution) and exposes one small section-level entry
+//! point per (architecture, scanned state): [`aarch64_downstream_flags_live`],
+//! [`aarch64_downstream_regs_live`], [`x86_downstream_flags_live`], and
+//! [`x86_downstream_regs_live`]. That wiring used to be copy-pasted in the CLI
+//! binary root, out of reach of library tests; consolidating it here keeps every
+//! downstream-liveness decision — and its soundness default — in one place.
 //!
 //! The per-instruction liveness primitives themselves live next door in
 //! [`super::live_out`] (`aarch64_reg_downstream_liveness`,
@@ -25,7 +31,12 @@
 use capstone::prelude::*;
 
 use super::live_out::DownstreamRegLiveness;
-use crate::isa::RegisterType;
+use crate::capstone_bridge::{ConvertOutcome, convert_capstone_op};
+use crate::elf_patcher::{AddressWindow, DetectedArch, ElfPatcher, TextSection};
+use crate::ir::{Instruction, Register};
+use crate::isa::x86::{X86Instruction, X86Register};
+use crate::isa::{FlagsAnalysis, RegisterType, X86_32, X86_64};
+use crate::parser::x86::x86_ir_from_mnemonic;
 use crate::semantics::live_out::RegisterSet;
 
 /// One decoded step of a downstream suffix scan — the caller's `decode` closure
@@ -204,10 +215,219 @@ where
     pin_all_remaining_live!();
 }
 
+// ---------------------------------------------------------------------------
+// Per-architecture wiring. The scanners above own the shared discipline; the
+// functions below supply the two knobs (Capstone→IR decode, per-instruction
+// rule), resolve the fall-through suffix from the ELF section, and expose one
+// small section-level entry point per (architecture, scanned state). These used
+// to live in the CLI binary root, out of reach of library tests; keeping them
+// here co-locates all downstream-liveness logic in one module.
+// ---------------------------------------------------------------------------
+
+/// Resolve the in-section fall-through suffix after an optimization window
+/// ending at `end_addr`. `None` means there is no analyzable suffix — the
+/// window already reaches the section end, or the bytes are unavailable — in
+/// which case downstream liveness is unknown and the caller keeps every
+/// candidate live (the conservative default). The section-level entry points
+/// below all funnel through here rather than repeating the suffix math.
+fn fall_through_suffix(
+    patcher: &ElfPatcher,
+    section: &TextSection,
+    end_addr: u64,
+) -> Option<Vec<u8>> {
+    let section_end = section.virtual_addr + section.size;
+    if end_addr >= section_end {
+        return None;
+    }
+    let suffix_window = AddressWindow {
+        start: end_addr,
+        end: section_end,
+    };
+    patcher.get_instructions_in_window(&suffix_window).ok()
+}
+
+/// Decode one AArch64 Capstone `(mnemonic, op_str)` pair into a downstream scan
+/// step, reusing the shared Capstone→IR bridge so the fall-through scan honors
+/// exactly the same supported-mnemonic set as the optimizer.
+fn aarch64_scan_step(mnemonic: &str, op_str: &str) -> ScanStep<Instruction> {
+    match convert_capstone_op(mnemonic, op_str) {
+        ConvertOutcome::Instruction(instr) => ScanStep::Decoded(instr),
+        ConvertOutcome::Skip => ScanStep::Skipped,
+        ConvertOutcome::Unsupported(_) => ScanStep::Opaque,
+    }
+}
+
+fn aarch64_downstream_flags_live_from_bytes(cs: &Capstone, bytes: &[u8], start_addr: u64) -> bool {
+    scan_flags_live(
+        cs,
+        bytes,
+        start_addr,
+        aarch64_scan_step,
+        |instr: &Instruction| {
+            super::live_out::flags_read_before_overwrite_after_window(std::slice::from_ref(instr))
+        },
+        |instr: &Instruction| instr.modifies_flags(),
+        |instr: &Instruction| instr.is_terminator(),
+    )
+}
+
+fn aarch64_downstream_regs_live_from_bytes(
+    cs: &Capstone,
+    bytes: &[u8],
+    start_addr: u64,
+    candidates: &RegisterSet<Register>,
+) -> RegisterSet<Register> {
+    scan_regs_live(
+        cs,
+        bytes,
+        start_addr,
+        candidates,
+        aarch64_scan_step,
+        |instr: &Instruction| instr.is_terminator(),
+        |reg: Register, instr: &Instruction| {
+            super::live_out::aarch64_reg_downstream_liveness(reg, std::slice::from_ref(instr))
+        },
+    )
+}
+
+/// Decode one x86 Capstone `(mnemonic, op_str)` pair into a downstream scan
+/// step. `nop` carries no observable state and is stepped over; anything the
+/// shared x86 IR does not model (including `call`/`ret`) is opaque and pins the
+/// remaining state live.
+fn x86_scan_step(mnemonic: &str, op_str: &str) -> ScanStep<X86Instruction> {
+    match x86_ir_from_mnemonic(mnemonic, op_str) {
+        Ok(Some(instr)) => ScanStep::Decoded(instr),
+        Ok(None) if mnemonic.eq_ignore_ascii_case("nop") => ScanStep::Skipped,
+        Ok(None) => ScanStep::Opaque,
+        Err(_) => ScanStep::Opaque,
+    }
+}
+
+fn x86_downstream_flags_live_from_bytes<I>(cs: &Capstone, bytes: &[u8], start_addr: u64) -> bool
+where
+    I: FlagsAnalysis<X86Instruction>,
+{
+    scan_flags_live(
+        cs,
+        bytes,
+        start_addr,
+        x86_scan_step,
+        |instr: &X86Instruction| <I as FlagsAnalysis<X86Instruction>>::reads_flags(instr),
+        |instr: &X86Instruction| <I as FlagsAnalysis<X86Instruction>>::modifies_flags(instr),
+        |instr: &X86Instruction| instr.is_terminator(),
+    )
+}
+
+fn x86_downstream_regs_live_from_bytes(
+    cs: &Capstone,
+    bytes: &[u8],
+    start_addr: u64,
+    candidates: &RegisterSet<X86Register>,
+) -> RegisterSet<X86Register> {
+    scan_regs_live(
+        cs,
+        bytes,
+        start_addr,
+        candidates,
+        x86_scan_step,
+        |instr: &X86Instruction| instr.is_terminator(),
+        |reg: X86Register, instr: &X86Instruction| {
+            super::live_out::x86_reg_downstream_liveness(reg, std::slice::from_ref(instr))
+        },
+    )
+}
+
+/// Whether the AArch64 condition flags (NZCV) may be read downstream of a window
+/// ending at `end_addr`. Returns the conservative `true` whenever the
+/// fall-through suffix is unavailable or the window already reaches the section
+/// end; otherwise scans the suffix and returns `false` only when a decoded
+/// instruction is proven to overwrite the flags before any read.
+pub fn aarch64_downstream_flags_live(
+    patcher: &ElfPatcher,
+    section: &TextSection,
+    end_addr: u64,
+    cs: &Capstone,
+) -> bool {
+    match fall_through_suffix(patcher, section, end_addr) {
+        Some(bytes) => aarch64_downstream_flags_live_from_bytes(cs, &bytes, end_addr),
+        None => true,
+    }
+}
+
+/// The subset of `candidates` (registers the AArch64 window writes) that are
+/// provably live downstream of the window. Returns every candidate live when the
+/// suffix is unavailable or the window already reaches the section end (issue
+/// #621's conservative default).
+pub fn aarch64_downstream_regs_live(
+    patcher: &ElfPatcher,
+    section: &TextSection,
+    end_addr: u64,
+    cs: &Capstone,
+    candidates: &RegisterSet<Register>,
+) -> RegisterSet<Register> {
+    match fall_through_suffix(patcher, section, end_addr) {
+        Some(bytes) => aarch64_downstream_regs_live_from_bytes(cs, &bytes, end_addr, candidates),
+        None => candidates.clone(),
+    }
+}
+
+/// Whether the x86 condition flags (EFLAGS) may be read downstream of a window
+/// ending at `end_addr`. Dispatches to the mode-specific flag analysis; an
+/// unavailable suffix and non-x86 `arch` values return the conservative `true`.
+pub fn x86_downstream_flags_live(
+    arch: DetectedArch,
+    patcher: &ElfPatcher,
+    section: &TextSection,
+    end_addr: u64,
+    cs: &Capstone,
+) -> bool {
+    let Some(bytes) = fall_through_suffix(patcher, section, end_addr) else {
+        return true;
+    };
+
+    match arch {
+        DetectedArch::X86_64 => {
+            x86_downstream_flags_live_from_bytes::<X86_64>(cs, &bytes, end_addr)
+        }
+        DetectedArch::X86_32 => {
+            x86_downstream_flags_live_from_bytes::<X86_32>(cs, &bytes, end_addr)
+        }
+        DetectedArch::Aarch64 => true,
+    }
+}
+
+/// The subset of `candidates` (registers the x86 window writes) that are provably
+/// live downstream. Register liveness is width-independent, so the
+/// mode-configured `cs` drives the correct x86-32/x86-64 disassembly; an
+/// unavailable suffix and non-x86 `arch` values return every candidate live.
+pub fn x86_downstream_regs_live(
+    arch: DetectedArch,
+    patcher: &ElfPatcher,
+    section: &TextSection,
+    end_addr: u64,
+    cs: &Capstone,
+    candidates: &RegisterSet<X86Register>,
+) -> RegisterSet<X86Register> {
+    let Some(bytes) = fall_through_suffix(patcher, section, end_addr) else {
+        return candidates.clone();
+    };
+
+    match arch {
+        DetectedArch::X86_64 | DetectedArch::X86_32 => {
+            x86_downstream_regs_live_from_bytes(cs, &bytes, end_addr, candidates)
+        }
+        DetectedArch::Aarch64 => candidates.clone(),
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::ir::Register;
+    use crate::assembler::AArch64Assembler;
+    use crate::assembler::x86::X86Assembler;
+    use crate::ir::{Condition, Instruction, LabelId, Operand, Register};
+    use crate::isa::x86::{X86Instruction, X86Register};
+    use crate::test_utils::{TempFile, build_minimal_aarch64_elf, build_minimal_x86_64_elf};
 
     /// AArch64 `NOP` (`0xD503201F`, little-endian). Decodes to exactly one
     /// instruction so the scripted `decode` closure is invoked once per copy.
@@ -412,5 +632,515 @@ mod tests {
             vec![step],
         );
         assert_eq!(sorted(&out), vec![Register::X0, Register::X2]);
+    }
+
+    // ---- Section-level entry points (the deep public interface) ----
+
+    fn assemble_aarch64_test_bytes(instructions: &[Instruction]) -> Vec<u8> {
+        AArch64Assembler::new()
+            .assemble_instructions(instructions, 0x1000)
+            .expect("test instruction should assemble")
+    }
+
+    fn assemble_x86_64_test_bytes(instructions: &[X86Instruction]) -> Vec<u8> {
+        X86Assembler::new_64()
+            .assemble_instructions(instructions)
+            .expect("test instruction should assemble")
+    }
+
+    fn aarch64_test_capstone() -> Capstone {
+        Capstone::new()
+            .arm64()
+            .mode(capstone::arch::arm64::ArchMode::Arm)
+            .detail(true)
+            .build()
+            .expect("test capstone should build")
+    }
+
+    fn x86_64_test_capstone() -> Capstone {
+        Capstone::new()
+            .x86()
+            .mode(capstone::arch::x86::ArchMode::Mode64)
+            .syntax(capstone::arch::x86::ArchSyntax::Intel)
+            .detail(true)
+            .build()
+            .expect("test capstone should build")
+    }
+
+    /// Materialize `elf` to a temp file and return its patcher plus the single
+    /// `.text` section. The `TempFile` is returned so the caller keeps the
+    /// backing file alive for the patcher's lifetime.
+    fn patcher_with_text(elf: &[u8]) -> (TempFile, ElfPatcher, TextSection) {
+        let file = TempFile::new_bytes("s11-downstream-section", "elf", elf);
+        let patcher = ElfPatcher::new(file.path()).expect("synthetic ELF should parse");
+        let section = patcher
+            .get_text_sections()
+            .expect("ELF should expose an executable section")
+            .into_iter()
+            .next()
+            .expect("minimal ELF should contain .text");
+        (file, patcher, section)
+    }
+
+    #[test]
+    fn aarch64_flags_live_defaults_conservative_at_section_end() {
+        // The window covers the whole section, so there is no analyzable
+        // fall-through suffix and the flags stay conservatively live.
+        let text = assemble_aarch64_test_bytes(&[Instruction::Add {
+            rd: Register::X1,
+            rn: Register::X2,
+            rm: Operand::Immediate(1),
+        }]);
+        let (_file, patcher, section) =
+            patcher_with_text(&build_minimal_aarch64_elf(&text, 0x1000));
+        let cs = aarch64_test_capstone();
+        let end = section.virtual_addr + section.size;
+        assert!(aarch64_downstream_flags_live(&patcher, &section, end, &cs));
+    }
+
+    #[test]
+    fn aarch64_flags_dead_when_suffix_proves_them_dead() {
+        // Window is the empty prefix; the whole section is the fall-through
+        // suffix. A lone non-flag ADD then the section ends, so nothing can
+        // read the flags: they are dead.
+        let text = assemble_aarch64_test_bytes(&[Instruction::Add {
+            rd: Register::X1,
+            rn: Register::X2,
+            rm: Operand::Immediate(1),
+        }]);
+        let (_file, patcher, section) =
+            patcher_with_text(&build_minimal_aarch64_elf(&text, 0x1000));
+        let cs = aarch64_test_capstone();
+        assert!(!aarch64_downstream_flags_live(
+            &patcher,
+            &section,
+            section.virtual_addr,
+            &cs
+        ));
+    }
+
+    #[test]
+    fn aarch64_regs_default_all_live_at_section_end() {
+        let text = assemble_aarch64_test_bytes(&[Instruction::MovReg {
+            rd: Register::X0,
+            rn: Register::X1,
+        }]);
+        let (_file, patcher, section) =
+            patcher_with_text(&build_minimal_aarch64_elf(&text, 0x1000));
+        let cs = aarch64_test_capstone();
+        let candidates = RegisterSet::from_registers(vec![Register::X0]);
+        let end = section.virtual_addr + section.size;
+        let live = aarch64_downstream_regs_live(&patcher, &section, end, &cs, &candidates);
+        assert!(live.contains(Register::X0));
+    }
+
+    #[test]
+    fn aarch64_regs_dead_when_suffix_fully_overwrites() {
+        // Suffix `mov x0, x1` fully overwrites X0 before any read → dead.
+        let text = assemble_aarch64_test_bytes(&[Instruction::MovReg {
+            rd: Register::X0,
+            rn: Register::X1,
+        }]);
+        let (_file, patcher, section) =
+            patcher_with_text(&build_minimal_aarch64_elf(&text, 0x1000));
+        let cs = aarch64_test_capstone();
+        let candidates = RegisterSet::from_registers(vec![Register::X0]);
+        let live = aarch64_downstream_regs_live(
+            &patcher,
+            &section,
+            section.virtual_addr,
+            &cs,
+            &candidates,
+        );
+        assert!(!live.contains(Register::X0));
+    }
+
+    #[test]
+    fn x86_flags_live_defaults_conservative_at_section_end() {
+        let text = assemble_x86_64_test_bytes(&[X86Instruction::MovImm {
+            rd: X86Register::RAX,
+            imm: 0,
+        }]);
+        let (_file, patcher, section) = patcher_with_text(&build_minimal_x86_64_elf(&text, 0x1000));
+        let cs = x86_64_test_capstone();
+        let end = section.virtual_addr + section.size;
+        assert!(x86_downstream_flags_live(
+            DetectedArch::X86_64,
+            &patcher,
+            &section,
+            end,
+            &cs
+        ));
+    }
+
+    #[test]
+    fn x86_flags_dead_when_suffix_proves_them_dead() {
+        // The whole section is the suffix: a lone non-flag `mov rax, 0` then the
+        // section ends, so the flags are dead.
+        let text = assemble_x86_64_test_bytes(&[X86Instruction::MovImm {
+            rd: X86Register::RAX,
+            imm: 0,
+        }]);
+        let (_file, patcher, section) = patcher_with_text(&build_minimal_x86_64_elf(&text, 0x1000));
+        let cs = x86_64_test_capstone();
+        assert!(!x86_downstream_flags_live(
+            DetectedArch::X86_64,
+            &patcher,
+            &section,
+            section.virtual_addr,
+            &cs
+        ));
+    }
+
+    #[test]
+    fn x86_regs_default_all_live_at_section_end() {
+        let text = assemble_x86_64_test_bytes(&[X86Instruction::MovImm {
+            rd: X86Register::RAX,
+            imm: 0,
+        }]);
+        let (_file, patcher, section) = patcher_with_text(&build_minimal_x86_64_elf(&text, 0x1000));
+        let cs = x86_64_test_capstone();
+        let candidates = RegisterSet::from_registers(vec![X86Register::RAX]);
+        let end = section.virtual_addr + section.size;
+        let live = x86_downstream_regs_live(
+            DetectedArch::X86_64,
+            &patcher,
+            &section,
+            end,
+            &cs,
+            &candidates,
+        );
+        assert!(live.contains(X86Register::RAX));
+    }
+
+    #[test]
+    fn x86_regs_dead_when_suffix_fully_overwrites() {
+        // Suffix `mov rax, 0` is a full-register write before any read → dead.
+        let text = assemble_x86_64_test_bytes(&[X86Instruction::MovImm {
+            rd: X86Register::RAX,
+            imm: 0,
+        }]);
+        let (_file, patcher, section) = patcher_with_text(&build_minimal_x86_64_elf(&text, 0x1000));
+        let cs = x86_64_test_capstone();
+        let candidates = RegisterSet::from_registers(vec![X86Register::RAX]);
+        let live = x86_downstream_regs_live(
+            DetectedArch::X86_64,
+            &patcher,
+            &section,
+            section.virtual_addr,
+            &cs,
+            &candidates,
+        );
+        assert!(!live.contains(X86Register::RAX));
+    }
+
+    // ---- Byte-level scanner characterization (migrated from the CLI binary) ----
+
+    #[test]
+    fn downstream_flags_live_scan_marks_dead_when_first_flag_event_writes() {
+        let bytes = assemble_aarch64_test_bytes(&[
+            Instruction::Cmp {
+                rn: Register::X0,
+                rm: Operand::Immediate(0),
+            },
+            Instruction::Csel {
+                rd: Register::X1,
+                rn: Register::X2,
+                rm: Register::X3,
+                cond: Condition::EQ,
+            },
+        ]);
+        let cs = aarch64_test_capstone();
+
+        assert!(!aarch64_downstream_flags_live_from_bytes(
+            &cs, &bytes, 0x1000
+        ));
+    }
+
+    #[test]
+    fn downstream_flags_live_scan_marks_live_when_first_flag_event_reads() {
+        let bytes = assemble_aarch64_test_bytes(&[Instruction::Csel {
+            rd: Register::X1,
+            rn: Register::X2,
+            rm: Register::X3,
+            cond: Condition::EQ,
+        }]);
+        let cs = aarch64_test_capstone();
+
+        assert!(aarch64_downstream_flags_live_from_bytes(
+            &cs, &bytes, 0x1000
+        ));
+    }
+
+    #[test]
+    fn downstream_flags_live_scan_marks_dead_for_known_non_flag_suffix() {
+        let bytes = assemble_aarch64_test_bytes(&[Instruction::Add {
+            rd: Register::X1,
+            rn: Register::X2,
+            rm: Operand::Immediate(1),
+        }]);
+        let cs = aarch64_test_capstone();
+
+        assert!(!aarch64_downstream_flags_live_from_bytes(
+            &cs, &bytes, 0x1000
+        ));
+    }
+
+    #[test]
+    fn downstream_flags_live_scan_is_conservative_for_unknown_context() {
+        let cs = aarch64_test_capstone();
+
+        assert!(aarch64_downstream_flags_live_from_bytes(&cs, &[], 0x1000));
+        assert!(aarch64_downstream_flags_live_from_bytes(
+            &cs,
+            &[0xff],
+            0x1000
+        ));
+        // LDR literal decodes in Capstone but is intentionally unsupported by
+        // the AArch64 optimization IR parser.
+        assert!(aarch64_downstream_flags_live_from_bytes(
+            &cs,
+            &[0x00, 0x00, 0x00, 0x58],
+            0x1000
+        ));
+    }
+
+    #[test]
+    fn downstream_flags_live_scan_is_conservative_for_unanalysed_branch() {
+        let bytes = assemble_aarch64_test_bytes(&[Instruction::B {
+            target: LabelId(0x1000),
+        }]);
+        let cs = aarch64_test_capstone();
+
+        assert!(aarch64_downstream_flags_live_from_bytes(
+            &cs, &bytes, 0x1000
+        ));
+    }
+
+    #[test]
+    fn x86_downstream_flags_live_scan_marks_live_when_first_flag_event_reads() {
+        use crate::isa::x86::X86Condition;
+
+        let bytes = assemble_x86_64_test_bytes(&[X86Instruction::Jcc {
+            cond: X86Condition::E,
+        }]);
+        let cs = x86_64_test_capstone();
+
+        assert!(x86_downstream_flags_live_from_bytes::<X86_64>(
+            &cs, &bytes, 0x1000
+        ));
+    }
+
+    #[test]
+    fn x86_downstream_flags_live_scan_marks_dead_when_first_flag_event_writes() {
+        let bytes = assemble_x86_64_test_bytes(&[
+            X86Instruction::CmpReg {
+                rn: X86Register::RAX,
+                rs: X86Register::RBX,
+            },
+            X86Instruction::MovImm {
+                rd: X86Register::RAX,
+                imm: 0,
+            },
+        ]);
+        let cs = x86_64_test_capstone();
+
+        assert!(!x86_downstream_flags_live_from_bytes::<X86_64>(
+            &cs, &bytes, 0x1000
+        ));
+    }
+
+    #[test]
+    fn x86_downstream_flags_live_scan_marks_dead_for_known_non_flag_suffix() {
+        let bytes = assemble_x86_64_test_bytes(&[X86Instruction::MovImm {
+            rd: X86Register::RAX,
+            imm: 0,
+        }]);
+        let cs = x86_64_test_capstone();
+
+        assert!(!x86_downstream_flags_live_from_bytes::<X86_64>(
+            &cs, &bytes, 0x1000
+        ));
+    }
+
+    #[test]
+    fn x86_downstream_flags_live_scan_is_conservative_for_unknown_context() {
+        let cs = x86_64_test_capstone();
+
+        assert!(x86_downstream_flags_live_from_bytes::<X86_64>(
+            &cs,
+            &[],
+            0x1000
+        ));
+        assert!(x86_downstream_flags_live_from_bytes::<X86_64>(
+            &cs,
+            &[0xff],
+            0x1000
+        ));
+        assert!(x86_downstream_flags_live_from_bytes::<X86_64>(
+            &cs,
+            &[0xc3],
+            0x1000
+        ));
+    }
+
+    fn x86_64_regset(regs: &[X86Register]) -> RegisterSet<X86Register> {
+        RegisterSet::from_registers(regs.to_vec())
+    }
+
+    fn aarch64_regset(regs: &[Register]) -> RegisterSet<Register> {
+        RegisterSet::from_registers(regs.to_vec())
+    }
+
+    #[test]
+    fn downstream_regs_live_scan_marks_dead_when_later_full_overwrite_precedes_any_read() {
+        // Window wrote X0. Suffix `mov x0, x1` fully overwrites x0 before any
+        // read, so X0 is dead/optimizable.
+        let bytes = assemble_aarch64_test_bytes(&[Instruction::MovReg {
+            rd: Register::X0,
+            rn: Register::X1,
+        }]);
+        let cs = aarch64_test_capstone();
+        let live = aarch64_downstream_regs_live_from_bytes(
+            &cs,
+            &bytes,
+            0x1000,
+            &aarch64_regset(&[Register::X0]),
+        );
+        assert!(
+            !live.contains(Register::X0),
+            "x0 fully overwritten before any read must be dropped from live-out"
+        );
+    }
+
+    #[test]
+    fn downstream_regs_live_scan_marks_live_when_read_before_overwrite() {
+        // Window wrote RAX. Suffix `add x2, x0, #1` reads x0 before any
+        // redefinition — x0 must stay live.
+        let bytes = assemble_aarch64_test_bytes(&[Instruction::Add {
+            rd: Register::X2,
+            rn: Register::X0,
+            rm: Operand::Immediate(1),
+        }]);
+        let cs = aarch64_test_capstone();
+        let live = aarch64_downstream_regs_live_from_bytes(
+            &cs,
+            &bytes,
+            0x1000,
+            &aarch64_regset(&[Register::X0]),
+        );
+        assert!(
+            live.contains(Register::X0),
+            "x0 read before any overwrite must stay live"
+        );
+    }
+
+    #[test]
+    fn downstream_regs_live_scan_conservative_for_unknown_context() {
+        let cs = aarch64_test_capstone();
+        let candidates = aarch64_regset(&[Register::X0, Register::X1]);
+
+        // Empty suffix → both candidates live.
+        let empty = aarch64_downstream_regs_live_from_bytes(&cs, &[], 0x1000, &candidates);
+        assert!(empty.contains(Register::X0) && empty.contains(Register::X1));
+
+        // Undisassemblable byte → live.
+        let garbage = aarch64_downstream_regs_live_from_bytes(&cs, &[0xff], 0x1000, &candidates);
+        assert!(garbage.contains(Register::X0) && garbage.contains(Register::X1));
+
+        // LDR-literal decodes in Capstone but is unsupported by the IR → live.
+        let unsupported = aarch64_downstream_regs_live_from_bytes(
+            &cs,
+            &[0x00, 0x00, 0x00, 0x58],
+            0x1000,
+            &candidates,
+        );
+        assert!(unsupported.contains(Register::X0) && unsupported.contains(Register::X1));
+    }
+
+    #[test]
+    fn downstream_regs_live_scan_marks_live_across_call_ret() {
+        let cs = aarch64_test_capstone();
+        let candidates = aarch64_regset(&[Register::X0, Register::X1]);
+
+        // `bl #0` is a call terminator → every window register may be
+        // observable across the ABI; keep them all live.
+        let bl_bytes = assemble_aarch64_test_bytes(&[Instruction::Bl {
+            target: LabelId(0x1000),
+        }]);
+        let across_call =
+            aarch64_downstream_regs_live_from_bytes(&cs, &bl_bytes, 0x1000, &candidates);
+        assert!(across_call.contains(Register::X0) && across_call.contains(Register::X1));
+
+        // `ret` is a return terminator → same ABI-observable rule.
+        let ret_bytes = assemble_aarch64_test_bytes(&[Instruction::Ret { rn: Register::X30 }]);
+        let across_ret =
+            aarch64_downstream_regs_live_from_bytes(&cs, &ret_bytes, 0x1000, &candidates);
+        assert!(across_ret.contains(Register::X0) && across_ret.contains(Register::X1));
+    }
+
+    #[test]
+    fn x86_partial_write_does_not_kill() {
+        // Window wrote RAX. Suffix `mov al, 0` leaves the rest of RAX intact,
+        // so the downstream scan must not treat it as a full-register kill.
+        let bytes = assemble_x86_64_test_bytes(&[X86Instruction::MovImm {
+            rd: X86Register::AL,
+            imm: 0,
+        }]);
+        let cs = x86_64_test_capstone();
+        let live = x86_downstream_regs_live_from_bytes(
+            &cs,
+            &bytes,
+            0x1000,
+            &x86_64_regset(&[X86Register::RAX]),
+        );
+        assert!(
+            live.contains(X86Register::RAX),
+            "an AL write preserves upper RAX bits, so RAX stays live"
+        );
+    }
+
+    #[test]
+    fn x86_downstream_regs_live_scan_marks_live_when_read_before_overwrite() {
+        // `add rbx, rax` reads rax before any redefinition → rax stays live.
+        let bytes = assemble_x86_64_test_bytes(&[X86Instruction::AddReg {
+            rd: X86Register::RBX,
+            rs: X86Register::RAX,
+        }]);
+        let cs = x86_64_test_capstone();
+        let live = x86_downstream_regs_live_from_bytes(
+            &cs,
+            &bytes,
+            0x1000,
+            &x86_64_regset(&[X86Register::RAX]),
+        );
+        assert!(live.contains(X86Register::RAX));
+    }
+
+    #[test]
+    fn x86_downstream_regs_live_scan_conservative_across_call_ret_and_unknown() {
+        let cs = x86_64_test_capstone();
+        let candidates = x86_64_regset(&[X86Register::RAX]);
+
+        // Empty suffix → live.
+        assert!(
+            x86_downstream_regs_live_from_bytes(&cs, &[], 0x1000, &candidates)
+                .contains(X86Register::RAX)
+        );
+        // `ret` (0xc3) is not modelled in the x86 IR → unsupported → live.
+        assert!(
+            x86_downstream_regs_live_from_bytes(&cs, &[0xc3], 0x1000, &candidates)
+                .contains(X86Register::RAX)
+        );
+        // `call rel32` (e8 00 00 00 00) is likewise not modelled → live.
+        assert!(
+            x86_downstream_regs_live_from_bytes(
+                &cs,
+                &[0xe8, 0x00, 0x00, 0x00, 0x00],
+                0x1000,
+                &candidates
+            )
+            .contains(X86Register::RAX)
+        );
     }
 }
